+++ gsops/README.md
