+++ gsops/_common.py
