+++ gsops/_exceptions.py
