+++ gsops/numpy/linalg.py
