+++ gsops/numpy/sparse.py
