+++ pyproject.toml
@@ -8,10 +8,10 @@ authors = [
     { name = "Luís F. Pereira", email = "luisfpereira@ucsb.edu" },
     { name = "Giulio Viganò", email = "g.vigano36@campus.unimib.it" },
 ]
-name = "gs"
+name = "gsops"
 dynamic = ["version"]
 readme = "README.rst"
-description = "Geomstats backend."
+description = "Geomstats backend in isolation."
 license = { file = "LICENSE.md" }
 classifiers = [
     "License :: OSI Approved :: MIT License",
@@ -36,13 +36,13 @@ autograd = ["autograd >= 1.3"]
 pytorch = ["torch >= 1.9.1"]
 
 [project.urls]
-repository = "http://github.com/luisfpereira/gs"
+repository = "http://github.com/luisfpereira/gsops"
 
 [tool.setuptools.dynamic]
-version = { attr = "gs.__version__" }
+version = { attr = "gsops.__version__" }
 
 [tool.setuptools.packages.find]
-include = ["gs", "gs.*"]
+include = ["gsops", "gsops.*"]
 
 [tool.pytest.ini_options]
 markers = [
@@ -64,12 +64,12 @@ extend-select = ["I", "D"]
 ignore = ["E731"]
 
 [tool.ruff.lint.isort]
-known-first-party = ["gs"]
+known-first-party = ["gsops"]
 
 [tool.ruff.lint.pydocstyle]
 convention = "numpy"
 
 [tool.ruff.lint.per-file-ignores]
 "*/__init__.py" = ["D104"]
-"gs/**/*" = ["D"]
+"gsops/**/*" = ["D"]
 "tests/*" = ["D"]
