+++ tests/data/against_einsum.py
@@ -1,6 +1,6 @@
 from geomstats.test.data import TestData
 
-import gs.backend as gs
+import gsops.backend as gs
 
 rand = gs.random.rand
