+++ tests/data/against_numpy.py
@@ -2,7 +2,7 @@ from geomstats.geometry.spd_matrices import SPDMatrices
 from geomstats.geometry.symmetric_matrices import SymmetricMatrices
 from geomstats.test.data import TestData
 
-import gs.backend as gs
+import gsops.backend as gs
 
 rand = gs.random.rand
